requirements.txt
numpy<2
fastapi
uvicorn
streamlit
requests
langchain
langchain-community
accelerate
transformers
sentence-transformers
huggingface-hub
pypdf
faiss-cpu
python-dotenv